+++ src/spmv_cpu/test_spmv_cpu.cpp
@@ -6,21 +6,21 @@
 #include <random>
 #include <vector>
 
-static std::vector<double> make_ones(int n) {
-    return std::vector<double>(n, 1.0);
+static std::vector<float> make_ones(int n) {
+    return std::vector<float>(n, 1.0f);
 }
 
-static std::vector<double> make_random(int n, unsigned seed = 42) {
+static std::vector<float> make_random(int n, unsigned seed = 42) {
     std::mt19937 rng(seed);
-    std::uniform_real_distribution<double> dist(0.0, 1.0);
-    std::vector<double> v(n);
+    std::uniform_real_distribution<float> dist(0.0f, 1.0f);
+    std::vector<float> v(n);
     for (auto& val : v) val = dist(rng);
     return v;
 }
 
 // Run SpMV and return result vector.
-static std::vector<double> run(const MtxCsr& A, const std::vector<double>& x) {
-    std::vector<double> y(A.num_rows, 0.0);
+static std::vector<float> run(const MtxCsr& A, const std::vector<float>& x) {
+    std::vector<float> y(A.num_rows, 0.0f);
     spmv_csr_cpu(A, x.data(), y.data());
     return y;
 }
@@ -47,8 +47,8 @@ int main(int argc, char** argv) {
         auto x = make_ones(A.num_cols);
         auto y1 = run(A, x);
         auto y2 = run(A, x);
-        double max_diff = compare_vectors(y1.data(), y2.data(), A.num_rows, 0.0);
-        bool pass = (max_diff == 0.0);
+        float max_diff = compare_vectors(y1.data(), y2.data(), A.num_rows, 0.0f);
+        bool pass = (max_diff == 0.0f);
         std::printf("[%s] Stage 1 (all-ones determinism): max_diff=%.3e\n",
                     pass ? "PASS" : "FAIL", max_diff);
         all_pass &= pass;
@@ -59,8 +59,8 @@ int main(int argc, char** argv) {
         auto x = make_random(A.num_cols, 42);
         auto y1 = run(A, x);
         auto y2 = run(A, x);
-        double max_diff = compare_vectors(y1.data(), y2.data(), A.num_rows, 0.0);
-        bool pass = (max_diff == 0.0);
+        float max_diff = compare_vectors(y1.data(), y2.data(), A.num_rows, 0.0f);
+        bool pass = (max_diff == 0.0f);
         std::printf("[%s] Stage 2 (random-seed-42 determinism): max_diff=%.3e\n",
                     pass ? "PASS" : "FAIL", max_diff);
         all_pass &= pass;
